Add data aligment---For 64bit differ 32bit.

+++ include/linux/highmem.h
@@ -4,6 +4,6 @@
 /*
  * This is virtual address that used in hihgmem.
  */
-#define KERNEL_RAM_VADDR   0xc0008000
-#define swapper_pg_dir     (KERNEL_RAM_VADDR - 0x00004000)
+#define KERNEL_RAM_VADDR   (unsigned long)0xc0008000
+#define swapper_pg_dir     (unsigned long)(KERNEL_RAM_VADDR - 0x00004000)
 #endif
